add to offer for options
